Fixing powerontimer
Due to the enhancement to 32 relays there are several changes of hard
coded limits required. Done for the powerontimer that switches offf a
relays after some time.

+++ sonoff/settings.h
@@ -45,7 +45,7 @@ struct SYSCFG {
   byte          button_restrict;       // Was message_format until 3.2.6a
   uint16_t      tele_period;
 
-  uint64_t      power;
+  uint32_t      power;
   uint8_t       ledstate;
   uint8_t       ex_switchmode;         // Not used since 3.9.21
 
@@ -155,8 +155,7 @@ struct RTCMEM {
   unsigned long hlw_kWhtoday;
 } rtcMem;
 
-// See issue https://github.com/esp8266/Arduino/issues/2913  
+// See issue https://github.com/esp8266/Arduino/issues/2913
 #ifdef USE_ADC_VCC
   ADC_MODE(ADC_VCC);                        // Set ADC input for Power Supply Voltage usage
 #endif
-
